+++ trilby-cli/src/Trilby/Command.hs
@@ -1,6 +1,8 @@
 module Trilby.Command where
 
 import Options.Applicative
+import Trilby.Infect.Options
+import Trilby.Infect.Options qualified as Infect
 import Trilby.Install.Options
 import Trilby.Install.Options qualified as Install
 import Trilby.Update.Options
@@ -10,13 +12,17 @@ import Prelude
 data Command m
     = Update (UpdateOpts m)
     | Install (InstallOpts m)
+    | Infect (InfectOpts m)
     deriving stock (Generic)
 
 parseCommand :: Parser (Command Maybe)
-parseCommand = hsubparser (parseUpdate <> parseInstall)
+parseCommand = hsubparser $ mconcat [parseUpdate, parseInstall, parseInfect]
+
+parseUpdate :: Mod CommandFields (Command Maybe)
+parseUpdate = command "update" $ info (Update <$> Update.parseOpts optional) (progDesc "Update Trilby")
 
 parseInstall :: Mod CommandFields (Command Maybe)
 parseInstall = command "install" $ info (Install <$> Install.parseOpts optional) (progDesc "Install Trilby")
 
-parseUpdate :: Mod CommandFields (Command Maybe)
-parseUpdate = command "update" $ info (Update <$> Update.parseOpts optional) (progDesc "Update Trilby")
+parseInfect :: Mod CommandFields (Command Maybe)
+parseInfect = command "infect" $ info (Infect <$> Infect.parseOpts optional) (progDesc "Infect with Trilby")