+++ trilby-cli/CHANGELOG.md
@@ -5,6 +5,8 @@
 * Print version when running verbosely
 * Use correct PATH when running in Nix sandbox
 * Update flakes recursively
+* Add Infect command
+
 ## 24.4.0 -- 2024-04-04
 
 * Set profile on update